Update build libraries and support for build 241
Updated the build libraries in the project. This includes upgrades to the Gradle version, adjusting the toolchain setup in the build configurations, and a refactor of Java command argument collection in the Gradle wrapper. In addition, support has been added for build 241 and updates were made to the Gradle actions used in the GitHub workflows.